edits: move around and create new structuring strategy for components and containers
