+++ client/src/components/Digest/style.css
@@ -1,0 +1,29 @@
+section.digest-posts {
+  margin: 0 auto;
+  padding: 0;
+}
+
+.digest-posts ul {
+  padding: 20px;
+  list-style: none;
+  display: flex;
+  flex-grow: 1;
+  flex-direction: row;
+  overflow-y: scroll;
+  -webkit-overflow-scrolling: touch;
+}
+
+.digest-posts ul li {
+  background: #ccc;
+  height: 145px;
+  width: 200px;
+  padding: 20px;
+  border-radius: 5px;
+  margin-right: 15px;
+}
+
+
+.digest-posts ul.party {
+  background-color: red;
+  padding: 20px;
+}
