+++ client/src/components/Header/Header.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-import './Header.css';
+import './style.css';
 
 export const Header = (props) => {
   return (
