+++ client/src/components/Header/style.css
@@ -5,6 +5,7 @@
   padding: 10px 0 15px;
   display: flex;
   flex-direction: row;
+  width: 100%
 
 }
 
@@ -13,6 +14,9 @@
   font-size: 1em;
   flex-grow: 1;
   margin: 0 10px;
+  display: inline-block;
+  width: calc(100% / 3);
+
 }
 
 .menu a {
