--- client/src/components/style.css
@@ -1,4 +1,0 @@
-.single {
-  height: 45px;
-  padding: 10px;
-}
